% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{BarplotPercentage}
\alias{BarplotPercentage}
\title{Barplot with percentage label}
\usage{
BarplotPercentage(dat, x, group, displayN = FALSE)
}
\arguments{
\item{dat}{dataframe}

\item{x}{x}

\item{group}{fill group}

\item{displayN}{If display sample number, default=FALSE}
}
\value{

}
\description{
Barplot with percentage label
}
