% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplotWithSCC}
\alias{scatterplotWithSCC}
\title{scatterplot with SCC
customized scatterplot}
\usage{
scatterplotWithSCC(dat, x, y, group = NULL, PCC = F, adj = NULL)
}
\arguments{
\item{dat}{dataframe, input}

\item{x}{string, lab for x}

\item{y}{string, lab for y}

\item{group}{string, lab for grouping, default NULL}

\item{PCC}{if permform ppcor::pcor.test}

\item{adj}{string, lab for confounder in pcor.test}
}
\value{

}
\description{
scatterplot with SCC
customized scatterplot
}
