% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{BarplotPercentage}
\alias{BarplotPercentage}
\title{Barplot with percentage label}
\usage{
BarplotPercentage(dat, x, group, yaxies = "p")
}
\arguments{
\item{dat}{dataframe}

\item{x}{x}

\item{group}{fill group}

\item{yaxies}{If display sample number (n) or percentage (p), default=p}
}
\value{

}
\description{
Barplot with percentage label
}
